fix videotrack framerate info

+++ interfaces/innerkits/native/media/demo/player/player_demo.cpp
@@ -306,7 +306,7 @@ int32_t PlayerDemo::GetVideoTrackInfo()
             iter->GetIntValue(std::string(PlayerKeys::PLAYER_BITRATE), bitrate);
             iter->GetIntValue(std::string(PlayerKeys::PLAYER_WIDTH), width);
             iter->GetIntValue(std::string(PlayerKeys::PLAYER_HEIGHT), height);
-            iter->GetIntValue(std::string(PlayerKeys::PLAYER_FRAMERATE), framerate);
+            iter->GetDoubleValue(std::string(PlayerKeys::PLAYER_FRAMERATE), framerate);
             iter->GetIntValue(std::string(PlayerKeys::PLAYER_TRACK_TYPE), type);
             iter->GetIntValue(std::string(PlayerKeys::PLAYER_TRACK_INDEX), index);
             cout << "mime: " << mime.c_str() << ", bitrate: " << bitrate << ", width: " << width << ", height: " <<
